Fixed bug with useLiveVue not available in SSR
