+++ assets/js/live_vue/index.ts
@@ -1,4 +1,3 @@
-import { type App, createSSRApp, createApp, Ref, Reactive } from 'vue'
-
 export { getHooks, initializeVueApp } from "./hooks"
 export { useLiveVue } from "./use"
+export { LiveVueOptions } from "./types"
